DM电机反转position取值变换更改。
